+++ pygeos/tests/test_predicates.py
@@ -44,7 +44,7 @@ BINARY_PREDICATES = (
     pygeos.covers,
     pygeos.covered_by,
     pytest.param(
-        partial(pygeos.distance_within, distance=1.0),
+        partial(pygeos.dwithin, distance=1.0),
         marks=pytest.mark.skipif(
             pygeos.geos_version < (3, 10, 0), reason="GEOS < 3.10"
         ),
@@ -117,18 +117,25 @@ def test_equals_exact_tolerance():
     assert pygeos.equals_exact(p1, p1).item() is True
     assert pygeos.equals_exact(p1, p2).item() is False
 
+    # an array of tolerances
+    actual = pygeos.equals_exact(p1, p2, tolerance=[0.05, 0.2, np.nan])
+    np.testing.assert_allclose(actual, [False, True, False])
+
 
 @pytest.mark.skipif(pygeos.geos_version < (3, 10, 0), reason="GEOS < 3.10")
-def test_distance_within():
-    # specifying tolerance
+def test_dwithin():
     p1 = pygeos.points(50, 4)
     p2 = pygeos.points(50.1, 4.1)
-    actual = pygeos.distance_within([p1, p2, None], p1, distance=0.05)
+    actual = pygeos.dwithin([p1, p2, None], p1, distance=0.05)
     np.testing.assert_allclose(actual, [True, False, False])
     assert actual.dtype == np.bool_
-    actual = pygeos.distance_within([p1, p2, None], p1, distance=0.2)
+    actual = pygeos.dwithin([p1, p2, None], p1, distance=0.2)
     np.testing.assert_allclose(actual, [True, True, False])
     assert actual.dtype == np.bool_
+
+    # an array of distances
+    actual = pygeos.dwithin(p1, p2, distance=[0.05, 0.2, np.nan])
+    np.testing.assert_allclose(actual, [False, True, False])
 
 
 @pytest.mark.parametrize(